ml_engine/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
numpy==2.2.0
scikit-learn==1.6.1
statsmodels==0.14.4
textblob==0.18.0
nltk==3.9.1
geopandas==1.0.1
shapely==2.0.6
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
python-dotenv==1.0.1
httpx==0.28.1
pydantic==2.10.4
